feat(getUserToken): allow calling with invoker
This allows to call getUserToken async.
It's useful when you want to be able to call it even when the library is
not fully loaded.

The goal for retrieving the user token is usally to add it as a header
on the searchClient. This enables Personalized results.

```ts
const searchClient = algoliasearch("APPLICATION_ID", "SEARCH_API_KEY");

// install search insights lib

aa('init', {...})
aa('getUserToken', _, (err, userToken) => {
  searchClient.setExtraHeader("X-Algolia-UserToken", userToken);
})

// search insights loads laters
```
This targets this JIRA ticket https://algolia.atlassian.net/browse/IFW-400