fix(processQueue): move callback on method level
I think this the way it should have been done in the first place.

Queue should pass down args agnostically.
Methods should decide what to do with the args, whether it's a callback or not.

This change results in a better separation of concerns, and will allow
implementing callback in other methods without any arity constraint.