fix(getUserToken): add integration tests
This test should have been added when the callback was implemented, but
it was only done as a unit test.

I'm about to change the way callbacks are passed to support detecting
when async XHR requests are done.

The pupose of this test is to secure the behaviour already documented
and released of `getUserToken`.