feat: restore anonymous user token on logout
This adds a mechanism to restore the previously assigned anonymous user token for logout scenarios.

It aims at providing a better API and functionality than the [current alternative explained in the documentation](https://www.algolia.com/doc/api-reference/api-methods/set-user-token/#examples).

When calling `setUserToken(undefined)` (on logout), we try to restore the anonymous user token found in the cookie, or generate another one.